requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-wtf>=1.2.0
requests>=2.31.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
flask-login
flask-bcrypt
flask-migrate